security/manager/ssl/tests/mochitest/bugs/test_generateCRMFRequest.html

<!DOCTYPE HTML>
<html>
<head>
  <title>crypto.generateCRMFRequest bugs</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="onWindowLoad()">
<script class="testbody" type="text/javascript">

  SimpleTest.waitForExplicitFinish();

  function onWindowLoad()
  {
    // Does it work at all?
    try {
      var crmfObject = crypto.generateCRMFRequest("CN=undefined", "regToken",
                                                  "authenticator", null, "",
                                                  512, null, "  rsa-ex   ",
                                                  1024, null, "\r\n\t rsa-sign\t");
      ok(true, "no exception thrown in generateCRMFRequest");
    } catch (e) {
      ok(false, "unexpected exception: " + e);
    }

    // bug 849553
    // This should fail because 8 is too small of a key size.
    try {
      var crmfObject = crypto.generateCRMFRequest("CN=undefined", "regToken",
                                                  "authenticator", null, "",
                                                  8, null, "rsa-ex",
                                                  1024, null, "rsa-sign");
      ok(false, "execution should not reach this line");
    } catch (e) {
      is(e.toString(), "Error: error:could not generate the key for algorithm rsa-ex", "expected exception");
    }
    // This should fail because 65536 is too large of a key size.
    try {
      var crmfObject = crypto.generateCRMFRequest("CN=undefined", "regToken",
                                                  "authenticator", null, "",
                                                  65536, null, "rsa-ex",
                                                  1024, null, "rsa-sign");
      ok(false, "execution should not reach this line");
    } catch (e) {
      is(e.toString(), "Error: error:could not generate the key for algorithm rsa-ex", "expected exception");
    }

    // bug 882865
    var o200 = document.documentElement;
    var o1 = crypto;
    try {
      o1.generateCRMFRequest("undefined", o200, 'X X X X X X X X X X X X X X X X X X X X X X X X X X X X X X X X X X X X X X', null, o1, 1404343237, Math.PI, []);
      ok(false, "execution should not reach this line");
    } catch (e) {
      // The 'key generation argument' in this case was an empty array,
      // which gets interpreted as an empty string.
      is(e.toString(), "Error: error:invalid key generation argument:", "expected exception");
    }

    // Test that only the first of repeated keygen parameters are used.
    try {
      var curveCrmfObject = crypto.generateCRMFRequest("CN=a", "a", "a", null,
                              "", 1024, "curve=secp521r1;curve=nistp384",
                              "ec-dual-use");
      ok(curveCrmfObject, "generateCRMFRequest succeeded");
      var curveRequest = curveCrmfObject.request;
      var curveBytes = atob(curveRequest.replace(/\r\n/g, ""));

      // nistp384 encoded in the request (as ASN1)
      var badIdentifier = [ 0x06, 0x05, 0x2B, 0x81, 0x04, 0x00, 0x22 ];
      ok(!findIdentifierInString(badIdentifier, curveBytes),
         "didn't find bad identifier in curve request");

      // secp512r1 encoded in the request (as ASN1)
      var goodIdentifier = [ 0x06, 0x05, 0x2B, 0x81, 0x04, 0x00, 0x23 ];
      ok(findIdentifierInString(goodIdentifier, curveBytes),
         "found good identifier in curve request");

      // The popcert=MII... values are base-64 encodings of self-signed
      // certificates. The key of the first one is a secp521r1 key, whereas
      // the second is nistp384.
      var popcertCrmfObject = crypto.generateCRMFRequest("CN=a", "a", "a",
                                null, "", 1024, "popcert=MIIBjjCB8aADAgECAgUAnVEmVTAJBgcqhkjOPQQBMAwxCjAIBgNVBAMTAWkwHhcNMTMxMTA2MjE1NDUxWhcNMTQwMjA2MjE1NDUxWjAMMQowCAYDVQQDEwFpMIGbMBAGByqGSM49AgEGBSuBBAAjA4GGAAQA5juM0J3Od2Ih4s0ZiTkzVkh96J4yO12/L71df3GWy/Ex3LaGcew/EucY5ZITEqNHn22+3pKf3pwL6GJ/zA/foJwBspiE42ITo2BHHXl2Uuf1QK70UH5gzaqf1Cc4lPv3ibOf3EAmHx0a8sdDxRlUN2+V38iYWnMmV1qf4jM15fsYEDMwCQYHKoZIzj0EAQOBjAAwgYgCQgDO34oQkVDNkwtto1OAEbFZgq1xP9aqc+Nt7vnOuEGTISdFXCjlhon7SysTejFhO8d8wG500NrlJEmaie9FJbmWpQJCARcVtnC0K+ilxz307GyuANTaLEd7vBJxTAho8l6xkibNoxY1D9+hcc6ECbK7PCtAaysZoAVx5XhJlsnFdJdDj3wG;popcert=MIIBRDCBy6ADAgECAgUAnVEotzAJBgcqhkjOPQQBMAwxCjAIBgNVBAMTAWkwHhcNMTMxMTA2MjIwMDExWhcNMTQwMjA2MjIwMDExWjAMMQowCAYDVQQDEwFpMHYwEAYHKoZIzj0CAQYFK4EEACIDYgAEXjFpZ9bodzikeN4C8p2mVj1Ia1t+8zIndSavQHmxaD3+kvhkt18+P20ZagfBOaVEQZdArZ6KxBeW9oYZqaNpqHLveGlKYi6u9z5FyozAx4MXzyLdfu+bzOLIsryKRnLFMAkGByqGSM49BAEDaQAwZgIxAJDawIJLQ5iZsJVC3vV1YEKsI2aNEicdZ3YTMp/zUy+64Z2/cjyyfa7d5m1xKLDBogIxANHOQoy/7DioCyWNDDzx5QK0M24dOURVWRXsxjAjrg4vDmV/fkVzwpUzIr5fMgXEyQ==", "ec-dual-use");
      ok(popcertCrmfObject, "generateCRMFRequest succeeded");
      var popcertRequest = popcertCrmfObject.request;
      var popcertBytes = atob(popcertRequest.replace(/\r\n/g, ""));
      ok(!findIdentifierInString(badIdentifier, popcertBytes),
         "didn't find bad identifier in popcert request");

      ok(findIdentifierInString(goodIdentifier, popcertBytes),
         "found good identifier in popcert request");
    } catch (e) {
      ok(false, "unexpected exception: " + e);
    }
    SimpleTest.finish();
  }

  function findIdentifierInString(identifier, str) {
    var matches = 0;
    for (var i = 0; i < str.length - identifier.length;
         i += (matches != 0 ? matches : 1)) {
      matches = 0;
      for (var j = 0; j < identifier.length; j++) {
        if (identifier[j] == str.charCodeAt(i + j)) {
          matches++;
        } else {
          break;
        }
      }
      if (matches == identifier.length) {
        return true;
      }
    }
    return false;
  }
</script>
</body>
</html>
